==> autocovid/jupyter notebook/1. github dataset preprocessing.ipynb <==
## [github](https://www.kaggle.com/kimjihoo/coronavirusdataset)에서 가져온 데이터 전처리

import sys
import os
from pathlib import Path

root = Path(os.getcwd()).parent.parent
sys.path.append(str(root))

import pandas as pd
import re

### PatientRoute.csv

patient_route = pd.read_csv(os.path.join(root, 'data', 'raw', 'github', 'PatientRoute.csv'))
patient_route.head()

### PatientInfo.csv

patient_info = pd.read_csv(os.path.join(root, 'data', 'raw', 'github', 'PatientInfo.csv'))
patient_info.head()

patient_info = patient_info[['patient_id', 'global_num', 'infection_case']]
patient_info.head()

### MergedRoute.csv

merged_route = pd.merge(patient_info, patient_route)
merged_route.head()

# latitude, longitude → grid 정보
def to_grid(lat, lon):
    row = 256
    column = 256
    seoul_start_lat = 37.698098
    seoul_start_lon = 126.799791
    seoul_end_lat = 37.472494
    seoul_end_lon = 127.142928
    lat_distance = seoul_end_lat - seoul_start_lat 
    lon_distance = seoul_end_lon - seoul_start_lon

    if seoul_end_lat > lat or lat > seoul_start_lat:
        return -1, -1
    if seoul_start_lon > lon or lon > seoul_end_lon:
        return -1, -1

    lat_distance2= lat - seoul_start_lat
    lon_distance2 = lon - seoul_start_lon
    row = (lat_distance2 * 255) / lat_distance
    col = (lon_distance2 * 255) / lon_distance

    return int(row), int(col)

merged_route['row'] = ""
merged_route['col'] = ""

for i in range(len(merged_route)):
    latitude = merged_route['latitude'].values[i]
    longitude = merged_route['longitude'].values[i]
    row, col = to_grid(latitude, longitude)
    merged_route.iloc[i, merged_route.columns.get_loc('row')] = row
    merged_route.iloc[i, merged_route.columns.get_loc('col')] = col
    
merged_route.head()

# 서울 아닌 확진자 방문지 제거, 필요 없는 columns 제거
merged_route = merged_route.loc[merged_route['row'] != -1]
merged_route = merged_route[['patient_id', 'city', 'infection_case', 'date', 'type', 'latitude', 'longitude', 'row', 'col']]
merged_route.head()

# city 형식에 맞게 변경
indices_to_del = []

for index, row in merged_route.iterrows():
    city = row['city']
    
    new_city = re.search(r'.*(?=-gu)', city)
    
    if new_city is None:
        indices_to_del.append(index)
        continue
        
    new_city = new_city.group().lower()
    merged_route.loc[index, 'city'] = new_city
    
for index in indices_to_del:
    merged_route = merged_route.drop(index=index)
    
merged_route = merged_route.reset_index(drop=True)
merged_route

# 파일로 저장
path = os.path.join(root, 'data', 'extracted', 'MergedRoute.csv')
merged_route.to_csv(path, encoding='utf-8-sig', index=False)
