==> autocovid/jupyter notebook/2-2. Merge Info.ipynb <==
import sys
import os
from pathlib import Path

root = Path(os.getcwd()).parent
sys.path.append(str(Path(os.getcwd()).parent))

from os import listdir
from os.path import isfile, join
import pandas as pd
import re
from datetime import datetime

route = pd.read_csv('aaa.csv', encoding='utf-8-sig', index_col=0)
route

patient = pd.read_csv(os.path.join(root, 'data', 'PatientInfo2.csv'), encoding='utf-8-sig', index_col=0)
patient

### 둘 다 확인 중인 경우

def check_check(reason):
    if '확인 중' in reason or '확인중' in reason:
        return True
    return False

# 둘 다 모두 확인 중인 경우 확인
numbers = []
for index1, row in patient.iterrows():
    number = row['환자 번호']
    route_rows = route.loc[route['patNo'] == number]
        
    count = 0
    for index2, route_row in route_rows.iterrows():
        if check_check(row['접촉력']) and check_check(route_row['reason']):
            count += 1
        
    if count == len(route_rows) and len(route_rows) != 0:
        numbers.append(number)
        
print(numbers)

# 삭제
for number in numbers:
    indices = route[route['patNo'] == number].index
    route = route.drop(indices, inplace=False)

route = route.reset_index(drop=True)
route

# 둘 다 확인 중이지만 다른 row에서는 그렇지 않은 경우
numbers = []
for index1, row in route.iterrows():
    number = row['patNo']
    patient_rows = patient.loc[patient['환자 번호'] == number]
    
    count = 0
    for index2, patient_row in patient_rows.iterrows():
        if check_check(patient_row['접촉력']) and check_check(row['reason']):
            count += 1
    
    if count > 0:
        numbers.append(number)
        
numbers = list(dict.fromkeys(numbers))
print(numbers)

# 다른 row에 있는 reason으로 변경
for number in numbers:
    unclear_route = route[route['patNo'] == number]
    reason = ''
    for row, elem in unclear_route.iterrows():
        if not check_check(elem['reason']):
            reason = elem['reason']
            
    indices = route[(route['patNo'] == number) & (route['reason'] == '확인 중')].index
    for index in indices:
        route.at[index, 'reason'] = reason
        
route

### 환자 정보에 등록되지 않았지만 다른 row에 reason이 있는 경우

# 환자 정보에 등록되지 않았지만 다른 row에 reason이 있는 경우
indices = route[route['reason'] == '확인 중'].index
indices_to_del = []
numbers = []
for index in indices:
    number = route.loc[index, 'patNo']
    reasons = patient[patient['환자 번호'] == number]['접촉력'].values
    
    if len(reasons) == 0:
        indices_to_del.append(index)
    else:
        numbers.append(number)
        route.at[index, 'reason'] = reasons[0]

numbers = list(dict.fromkeys(numbers))
print(numbers)
        
# route 모두 확인 중이고 환자 정보에도 등록되지 않은 경우 삭제
route = route.drop(indices_to_del, inplace=False)
route = route.reset_index(drop=True)
route

### '해외 접촉 추정' → 해외 접촉

indices = route[route['reason'].str.contains('해외 접촉|해외접촉', na=False)].index
for index in indices:
    route.at[indices, 'reason'] = '해외 접촉'
    
route[route['reason'].str.contains('해외 접촉', na=False)]

### 저장

path = os.path.join(root, 'data', 'modified_route.csv')
route.to_csv(path, encoding='utf-8-sig', index=False)
