==> autocovid/jupyter notebook/2-3. Geo.ipynb <==
import sys
import os
from pathlib import Path

root = Path(os.getcwd()).parent
sys.path.append(str(Path(os.getcwd()).parent))

from api import GeoCoader
import pandas as pd
import re
from datetime import datetime

path = os.path.join(root, 'data', 'modified_route.csv')
route = pd.read_csv(path, encoding='utf-8-sig')
route

### '소독'으로 시작하는 route 제외

indices = route[route['route'].str.contains('소독완료|소독 완료', na=False)].index
indices_to_del = []
for index in indices:
    mask = re.search("^소독", route.iloc[index]['route'])
    if mask is not None:
        indices_to_del.append(index)
        print(index, route.iloc[index]['route'])

route = route.drop(indices_to_del, inplace=False)
route = route.reset_index(drop=True)
route

### * 이후 제외

indices_to_del = []
for index, row in route.iterrows():
    place = row['route']
    
    # '*' 이후
    regex_searched = re.search('\*', place)
    if regex_searched is None:
        continue
    
    # '*' 제거 후 공백인 경우
    new_place = place[:regex_searched.span()[0]]
    space_searched = re.search('^\s*$\s*', new_place)
    if space_searched is not None:
        indices_to_del.append(index)
    else:
        route.at[index, 'route'] = new_place

# 공백인 경우 삭제
route = route.drop(indices_to_del, inplace=False)
route = route.reset_index(drop=True)
route

### use API

### name, type, latitude, longitude column 추가
route['type'] = ''
route['latitude'] = 0.0
route['longitude'] = 0.0
route

geo_api = GeoCoader()

print('size: %d' % len(route.index))
indices_to_del = []
for index, row in route.iterrows():
    if index % 100 == 0:
        print(index)
    
    information = geo_api.get_information(row['route'])
    if information.empty:
        indices_to_del.append(index)
        continue
    route.at[index, 'latitude'] = information['lat'].values[0]
    route.at[index, 'longitude'] = information['lng'].values[0]
    route.at[index, 'type'] = information['type'].values[0]
    
route = route.drop(indices_to_del, inplace=False)
route = route.reset_index(drop=True)
route

path = os.path.join(root, 'data', 'modified_route2.csv')
route.to_csv(path, encoding='utf-8-sig', index=False)
