==> autocovid/jupyter notebook/3-1. corona status.ipynb <==
## 서울특별시 코로나19 현황
https://www.seoul.go.kr/coronaV/coronaStatus.do

import sys
import os
from pathlib import Path

root = Path(os.getcwd()).parent.parent
sys.path.append(str(root))

from bs4 import BeautifulSoup
from src.preprocess.parser import StatusParser
import requests
import pandas as pd

url = 'https://www.seoul.go.kr/coronaV/coronaStatus.do'
table_class = 'tstyle05 tstyleP status-datatable datatable-multi-row'

### 사이트에서 현황 가져오고 저장

parser = StatusParser(url, table_class)
html_path = os.path.join(root, 'data', 'raw', 'status', 'corona_status.html')

content = parser.request_status()
parser.save_html(html_path, content)

### 파일에 저장된 현황 가져와서 파싱 후 저장

parser = StatusParser(url, table_class)
content_path = os.path.join(root, 'data', 'raw', 'status', 'corona_status.html')
content = parser.read_html(content_path)

table = parser.extract_table(content)
patient_info = parser.extract_info(table)
info_path = os.path.join(root, 'data', 'extracted', 'corona_status.csv')
parser.save_info(info_path, patient_info)
