==> autocovid/jupyter notebook/4-3. merge all dataset.ipynb <==
## 데이터셋 병합
- 병합 데이터셋
  - `data\extracted\MergedRoute.csv`
  - `data\extracted\modified_route.csv`
  - `data\extracted\merged_route_check_04_03.csv`
- 최종 데이터 저장 경로
  - `data\extracted\merged_route_final.csv`

import sys
import os

from os import listdir
from os.path import isfile, join
from pathlib import Path

root = Path(os.getcwd()).parent.parent
sys.path.append(str(root))

import pandas as pd
import numpy as np
import re

### `MergedRoute.csv` 전처리 더 필요한 부분

# merged_df 확인
merged_path = join(root, 'data', 'extracted', 'MergedRoute.csv')
merged_df = pd.read_csv(merged_path)
merged_df.head()

# 중복되는 type 줄이기 위한 type.csv
type_path = join(root, 'data', 'extracted', 'type.csv')
normal_type_df = pd.read_csv(type_path)

normal_type_df.head()

# 중복된 type column value 변경
indices_to_del = []

for index, row in merged_df.iterrows():
    location_type = row['type']
    
    normal_row = normal_type_df[normal_type_df['type'] == location_type]
    if normal_row.empty:
        indices_to_del.append(index)
        continue
    else:
        normal_row = normal_row.iloc[0]
        
    if not pd.isnull(normal_row['to_replace']):
        merged_df.loc[index, 'type'] = normal_row['to_replace']
        
for index in indices_to_del:
    merged_df = modified_route.drop(index=index)
    
merged_df = merged_df.reset_index(drop=True)
merged_df.head()

### `modified_route.csv` 전처리 더 필요한 부분

# modified_df 확인
modified_path = join(root, 'data', 'extracted', 'modified_route.csv')
modified_df = pd.read_csv(modified_path)
modified_df.head()

# city 구하기 위해 status file 확인
status_path = join(root, 'data', 'extracted', 'corona_status.csv')
status_df = pd.read_csv(status_path)
status_df.head()

# modified_df의 patNo column과 status_df의 환자 번호 column 비교해서 거주지 복사
indices_to_del = []

for index, row in modified_df.iterrows():
    patient_id = row['patNo']
    
    patient_in_status = status_df[status_df['환자 번호'] == patient_id]
    if patient_in_status.empty:
        indices_to_del.append(index)
        continue
    
    modified_df.loc[index, 'city'] = patient_in_status.iloc[0]['거주지']
    
for index in indices_to_del:
    modified_df = modified_df.drop(index=index)
    
modified_df = modified_df.reset_index(drop=True)
modified_df.head()

# 한국어-영어 구
names_path = join(root, 'data', 'raw', 'names.csv')
names_df = pd.read_csv(names_path)

names_df.head()

# 한국어로 된 city 영어로 바꾸기
for index, row in modified_df.iterrows():
    city = row['city']
    if city[-1] == '구': city = city[0:len(city) - 1]
    
    name_row = names_df[names_df['korean'] == city]
    if name_row.empty:
        new_city = 'other'
    else:
        new_city = name_row.iloc[0]['english']
        
    modified_df.loc[index, 'city'] = new_city
    
modified_df.head()

# 필요한 column만 따로 빼내고, 이름 규격대로 변경
columns = ['date', 'patNo', 'reason', 'type', 'latitude', 'longitude', 'city']
modified_df = modified_df[columns]

new_columns = {'patNo': 'patient_id', 'reason': 'infection_case'}
modified_df.rename(columns=new_columns, inplace=True)
modified_df.head()

# 중복되는 type 줄이기 위한 type.csv
type_path = join(root, 'data', 'extracted', 'type.csv')
normal_type_df = pd.read_csv(type_path)

normal_type_df.head()

# 중복된 type column value 변경
indices_to_del = []

for index, row in modified_df.iterrows():
    location_type = row['type']
    
    normal_row = normal_type_df[normal_type_df['type'] == location_type]
    if normal_row.empty:
        indices_to_del.append(index)
        continue
    else:
        normal_row = normal_row.iloc[0]
        
    if not pd.isnull(normal_row['to_replace']):
        modified_df.loc[index, 'type'] = normal_row['to_replace']
        
for index in indices_to_del:
    modified_df = modified_route.drop(index=index)
    
modified_df = modified_df.reset_index(drop=True)
modified_df.head()

### `merged_route_check_04_03.csv` 전처리 더 필요한 부분

# last_df 확인
last_path = join(root, 'data', 'extracted', 'merged_route_check_04_03.csv')
last_df = pd.read_csv(last_path)
last_df.head()

# 한글 구 영어로 바꾸기 위한 names
names_path = join(root, 'data', 'raw', 'names.csv')
names_df = pd.read_csv(names_path)
names_df.head()

# from, patient_id 내용 확인하고 city column 추가
for index, row in last_df.iterrows():
    region = row['from']
    patient_id = row['patient_id']
    
    # patient_id가 숫자로만 이루어진 경우
    if re.search(r'^\d+$', patient_id) is not None:
        last_df.loc[index, 'city'] = region
    else:
        kor_region = re.search(r'[가-힣]+(?=\d+)', patient_id).group()
        name_row = names_df[names_df['korean'] == kor_region].iloc[0]
        eng_region = name_row['english']
        last_df.loc[index, 'city'] = eng_region
        
last_df.head()

# 한국어-영어 location_type 변경 위한 kor_type_df 확인
kor_type_path = join(root, 'data', 'extracted', 'kor_type.csv')
kor_type_df = pd.read_csv(kor_type_path)
kor_type_df.head()

# last_df에서 한국어로 된 location_type 영어로 변경
for index, row in last_df.iterrows():
    type_row = kor_type_df.loc[kor_type_df['type'] == row['location_type']]
    last_df.loc[index, 'type'] = type_row.iloc[0]['eng_type']
    
last_df.head()

# 필요한 column만 따로 빼내고, 이름 규격대로 변경
columns = ['date', 'reason', 'patient_id', 'lat', 'lng', 'type', 'city']
last_df = last_df[columns]

new_columns = {'reason': 'infection_case', 'lat': 'latitude', 'lng': 'longitude'}
last_df.rename(columns=new_columns, inplace=True)
last_df.head()

### merged_df, modified_df, last_df 병합

final_df = pd.concat([merged_df, modified_df, last_df], axis=0, ignore_index=True)
final_df = final_df.reset_index(drop=True)
final_df.head()

# row, col 없는 경우 값 채워놓기
def to_grid(lat, lon):
    row = 256
    column = 256
    seoul_start_lat = 37.698098
    seoul_start_lon = 126.799791
    seoul_end_lat = 37.472494
    seoul_end_lon = 127.142928
    lat_distance = seoul_end_lat - seoul_start_lat 
    lon_distance = seoul_end_lon - seoul_start_lon

    if seoul_end_lat > lat or lat > seoul_start_lat:
        return -1, -1
    if seoul_start_lon > lon or lon > seoul_end_lon:
        return -1, -1

    lat_distance2= lat - seoul_start_lat
    lon_distance2 = lon - seoul_start_lon
    row = (lat_distance2 * 255) / lat_distance
    col = (lon_distance2 * 255) / lon_distance

    return int(row), int(col)

for index, row in final_df.iterrows():
    if not pd.isnull(row['row']): continue
        
    latitude = row['latitude']
    longitude = row['longitude']
    row_value, col_value = to_grid(latitude, longitude)
    final_df.loc[index, 'row'] = row_value
    final_df.loc[index, 'col'] = col_value
    
final_df.head()

# infection_case column values 정규화하기 위한 파일
infection_path = join(root, 'data', 'extracted', 'infection_case.csv')
infection_df = pd.read_csv(infection_path)
infection_df.head()

# infection_case 정리
count = 0
indices_to_del = []

for index, row in final_df.iterrows():
    infection_case = row['infection_case']
    if pd.isnull(infection_case):
        indices_to_del.append(index)
        continue
    
    infection_row = infection_df.loc[infection_df['infection_case'] == infection_case].iloc[0]
    
    if pd.isnull(infection_row['to_replace']): continue
    final_df.loc[index, 'infection_case'] = infection_row['to_replace']
    count += 1
    
for index in indices_to_del:
    final_df = final_df.drop(index=index)
    
final_df = final_df.reset_index(drop=True)

print('%d rows are chaned' % count)
final_df.head()

# row, col 범위 잘못된 부분 삭제
indices_to_del = []

for index, route in final_df.iterrows():
    row = route['row']
    col = route['col']
    
    if pd.isnull(row) or pd.isnull(col):
        print('null')
        indices_to_del.append(index)
        continue
    if row >= 256 or col >= 256:
        indices_to_del.append(index)
        continue
    if row < 0 or col < 0:
        indices_to_del.append(index)
        continue

for index in indices_to_del:
    final_df = final_df.drop(index=index)
    
print('len after delete: %d' % len(final_df.index))

# date column 정규화
for index, row in final_df.iterrows():
    date_str = row['date']
    new_date = re.search(r'\d+-\d+-\d+', date_str).group()
    if date_str == new_date: continue
    final_df.loc[index, 'date'] = new_date

# 날짜 순으로 데이터 정렬
final_df = final_df.sort_values(by='date')
final_df = final_df.reset_index(drop=True)
final_df

# 저장
final_path = join(root, 'data', 'extracted', 'merged_route_final.csv')
final_df.to_csv(final_path, encoding='utf-8-sig', index=False)
