==> autocovid/jupyter notebook/5. column values to csv.ipynb <==
## Column values to csv
- `data\extracted\merged_route_final`에서 city, infection_case. type column values 파일로 따로 저장
- 저장 목록
  - `data\checklist\city_column.csv
  - `data\checklist\infection_case_column.csv
  - `data\checklist\type_case_column.csv

import sys
import os

from os import listdir
from os.path import isfile, join
from pathlib import Path

root = Path(os.getcwd()).parent.parent
sys.path.append(str(root))

import pandas as pd

final_path = join(root, 'data', 'extracted', 'merged_route_final.csv')
final_df = pd.read_csv(final_path)
final_df.head()

final_df.columns

city_list = final_df['city'].drop_duplicates().tolist()
len(city_list)

case_list = final_df['infection_case'].drop_duplicates().tolist()
len(case_list)

type_list = final_df['type'].drop_duplicates().tolist()
len(type_list)

city_path = join(root, 'data', 'checklist', 'city_column.csv')
city_df = pd.DataFrame(data={'city': city_list})
city_df.to_csv(city_path, encoding='utf-8-sig',index=False)

case_path = join(root, 'data', 'checklist', 'infection_case_column.csv')
case_df = pd.DataFrame(data={'infection_case': case_list})
case_df.to_csv(case_path, encoding='utf-8-sig',index=False)

type_path = join(root, 'data', 'checklist', 'type_column.csv')
type_df = pd.DataFrame(data={'type': type_list})
type_df.to_csv(type_path, encoding='utf-8-sig',index=False)
